TERMINAL TEXTURE
block id 2 has one of its faces (and it can ve specified durring runtime)
read from a texture that is in a format that can be written to from a 6502 emulator.

woot!

the only issue is its currently set up to be REALLY FREAKING SLOW in terms of fps.
its also nly set up to handly one texture.
will fox soon.