fixed memory fragging / leak
I still do not understand what was wrong with the previous code in
chunk.c... specifically with the reload functiuon. I found the
incorrect if statement before the free in the worldgen function... but
even with that fixed the pproblem niether got worse nor better. Now
chunks are simply overwritten and never recalloced. This will also
prevent long term memory fragmentation... so its better anyways.