cleaning chunk
many functions removed from chunk.h, and they have been ordered
logically.

all functions in the header are now in the c file

chunks and their data are now only allocated once on load... then
overwritten in every function.

the load functions return a pointer instead of filling a pointer be to
be consistant with malloc and the like... plus, the information about
wether or not the chunk was loaded from disk should be abstracted away
anyways (or at least in my opnion).

all refrences have been updated (only world.c)