Uncompress chunks under load
When a chunk has a lot of block updates flowing through it,
temporarily uncompress that chunk for faster reading and writing.

This is only useful when the same blocks keep geting updated (like
with water flowing), because all of the blocks still have to get pup
back into the octree eventually.